refactor size_t

+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: tlakchai <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/28 13:13:31 by tlakchai          #+#    #+#             */
-/*   Updated: 2023/08/30 14:27:49 by tlakchai         ###   ########.fr       */
+/*   Updated: 2023/08/30 18:07:07 by tlakchai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (!s || !f)
